Take into account an input tape alingment issue wrt context-checking.
When entering a right context, the checking proceeds from the current location
in the input tape, so it should not be incremented by tape_step. For left
contexts (LC and NLC), however, checking begins at the previous location,
so we need to increment right away. This will cause problems when combining
LC and RC contexts with OR...