This fixes the mysterious 24 + n*16 bug by padding for no apparent reason.
A bug has been present in get_quoted() / parse_quoted() for a while in which
a segfault occurs when the quoted string is a particular number of bytes
in length. I still don't understand what exactly happens, but the bad
accesses are always one byte too far (and the string appears to come
through correctly), so for the time being it's sufficient to allocate
strlen(quoted) + 1 bytes for the return value.