Large speedup in conversion to optimized-lookup format
In some cases this speeds up conversion by up to 80%. This involved
a) making the code a bit more convoluted by using multiple data structures
where there used to be just one
b) omitting some space-oriented optimizations that appear to be usually
very minor and occasionally even harmful

While I've tested this with the major transducers I've found, it's not
unthinkable that refactoring such hairy code as this has introduced bugs.
No new functionality is introduced so it's ok to roll this back if it causes
problems.